requirement.txt
fastapi==0.128.5
uvicorn==0.40.0
torch==2.4.1
torchvision==0.19.1
torchaudio==2.4.1
pillow
numpy
scikit-learn
python-multipart
